split google page and emailPage
