+++ package.json
@@ -15,9 +15,13 @@
         "@wdio/mocha-framework": "^8.2.3",
         "@wdio/spec-reporter": "^8.1.2",
         "chromedriver": "^109.0.0",
-        "wdio-chromedriver-service": "^8.0.1"
+        "wdio-chromedriver-service": "^8.0.1",
+        "wdio-electron-service": "^4.0.0"
     },
     "scripts": {
         "wdio": "wdio run wdio.conf.js"
+    },
+    "dependencies": {
+        "-": "^0.0.1"
     }
 }
