+++ pom/googlePageCheck.js
