+++ wdio.conf.js
@@ -6,7 +6,6 @@ exports.config = {
     // WebdriverIO supports running e2e tests as well as unit and component tests.
     runner: 'local',
     screenshotPath: './test/specs/screenshots',
-
     //
     // ==================
     // Specify Test Files
